# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PayMate - AUD on Solana API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Contact {
  id: string;
  name: string;
  walletAddress: string;
  email?: string | null;
  note?: string | null;
  createdAt: string;
}

export interface CreateContactBody {
  name: string;
  walletAddress: string;
  email?: string | null;
  note?: string | null;
}

export type InvoiceStatus = (typeof InvoiceStatus)[keyof typeof InvoiceStatus];

export const InvoiceStatus = {
  draft: "draft",
  sent: "sent",
  paid: "paid",
  overdue: "overdue",
} as const;

export interface Invoice {
  id: string;
  title: string;
  recipientName: string;
  recipientWallet: string;
  recipientEmail?: string | null;
  amountAudd: number;
  dueDate?: string | null;
  status: InvoiceStatus;
  txSignature?: string | null;
  note?: string | null;
  createdAt: string;
  paidAt?: string | null;
}

export type CreateInvoiceBodyStatus =
  (typeof CreateInvoiceBodyStatus)[keyof typeof CreateInvoiceBodyStatus];

export const CreateInvoiceBodyStatus = {
  draft: "draft",
  sent: "sent",
  paid: "paid",
  overdue: "overdue",
} as const;

export interface CreateInvoiceBody {
  title: string;
  recipientName: string;
  recipientWallet: string;
  recipientEmail?: string | null;
  amountAudd: number;
  dueDate?: string | null;
  note?: string | null;
  status: CreateInvoiceBodyStatus;
}

export interface MarkPaidBody {
  txSignature?: string | null;
}

export interface PaymentLink {
  id: string;
  label: string;
  amountAudd?: number | null;
  walletAddress: string;
  slug: string;
  active: boolean;
  totalReceived: number;
  paymentCount: number;
  createdAt: string;
}

export interface CreatePaymentLinkBody {
  label: string;
  amountAudd?: number | null;
  walletAddress: string;
}

export type PaymentRequestStatus =
  (typeof PaymentRequestStatus)[keyof typeof PaymentRequestStatus];

export const PaymentRequestStatus = {
  pending: "pending",
  paid: "paid",
  cancelled: "cancelled",
} as const;

export interface PaymentRequest {
  id: string;
  toName: string;
  toWallet: string;
  amountAudd: number;
  note?: string | null;
  status: PaymentRequestStatus;
  createdAt: string;
  paidAt?: string | null;
}

export interface CreatePaymentRequestBody {
  toName: string;
  toWallet: string;
  amountAudd: number;
  note?: string | null;
}

export type RecurringPaymentFrequency =
  (typeof RecurringPaymentFrequency)[keyof typeof RecurringPaymentFrequency];

export const RecurringPaymentFrequency = {
  weekly: "weekly",
  monthly: "monthly",
} as const;

export interface RecurringPayment {
  id: string;
  label: string;
  recipientName: string;
  recipientWallet: string;
  amountAudd: number;
  frequency: RecurringPaymentFrequency;
  nextRunAt: string;
  active: boolean;
  createdAt: string;
}

export type CreateRecurringPaymentBodyFrequency =
  (typeof CreateRecurringPaymentBodyFrequency)[keyof typeof CreateRecurringPaymentBodyFrequency];

export const CreateRecurringPaymentBodyFrequency = {
  weekly: "weekly",
  monthly: "monthly",
} as const;

export interface CreateRecurringPaymentBody {
  label: string;
  recipientName: string;
  recipientWallet: string;
  amountAudd: number;
  frequency: CreateRecurringPaymentBodyFrequency;
}

export interface SplitParticipant {
  name: string;
  walletAddress: string;
  shareAudd: number;
  settled: boolean;
}

export interface Split {
  id: string;
  title: string;
  totalAudd: number;
  participants: SplitParticipant[];
  createdAt: string;
}

export interface CreateSplitParticipant {
  name: string;
  walletAddress: string;
  shareAudd: number;
}

export interface CreateSplitBody {
  title: string;
  totalAudd: number;
  participants: CreateSplitParticipant[];
}

export interface SettleParticipantBody {
  walletAddress: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  send: "send",
  receive: "receive",
  invoice: "invoice",
  split: "split",
  payment_link: "payment_link",
  recurring: "recurring",
} as const;

export interface Transaction {
  id: string;
  type: TransactionType;
  counterpartyName?: string | null;
  counterpartyWallet?: string | null;
  amountAudd: number;
  txSignature?: string | null;
  note?: string | null;
  createdAt: string;
}

export type CreateTransactionBodyType =
  (typeof CreateTransactionBodyType)[keyof typeof CreateTransactionBodyType];

export const CreateTransactionBodyType = {
  send: "send",
  receive: "receive",
} as const;

export interface CreateTransactionBody {
  type: CreateTransactionBodyType;
  counterpartyName?: string | null;
  counterpartyWallet?: string | null;
  amountAudd: number;
  txSignature?: string | null;
  note?: string | null;
}

export interface TransactionCsvRow {
  date: string;
  type: string;
  counterparty: string;
  amountAudd: number;
  txSignature: string;
  note: string;
}

export interface DashboardSummary {
  balanceAudd: number;
  balanceAud: number;
  balanceUsd: number;
  totalSentAudd: number;
  totalReceivedAudd: number;
  pendingInvoicesCount: number;
  pendingInvoicesAudd: number;
  activePaymentLinks: number;
  activeRecurring: number;
  recentTransactions: Transaction[];
}

export interface ExchangeRates {
  AUDD_AUD: number;
  AUDD_USD: number;
  AUDD_SOL: number;
  updatedAt: string;
}