# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PayMate - AUD on Solana API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payment-requests:
    get:
      operationId: listPaymentRequests
      tags: [payment-requests]
      summary: List all payment requests
      responses:
        "200":
          description: List of payment requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentRequest"
    post:
      operationId: createPaymentRequest
      tags: [payment-requests]
      summary: Create a payment request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequestBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRequest"
  /payment-requests/{id}:
    get:
      operationId: getPaymentRequest
      tags: [payment-requests]
      summary: Get a payment request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRequest"
    delete:
      operationId: deletePaymentRequest
      tags: [payment-requests]
      summary: Delete a payment request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    PaymentRequest:
      type: object
      properties:
        id:
          type: string
        toName:
          type: string
        toWallet:
          type: string
        amountAudd:
          type: number
        note:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, paid, cancelled]
        createdAt:
          type: string
        paidAt:
          type: string
          nullable: true
      required: [id, toName, toWallet, amountAudd, status, createdAt]
    CreatePaymentRequestBody:
      type: object
      properties:
        toName:
          type: string
        toWallet:
          type: string
        amountAudd:
          type: number
        note:
          type: string
          nullable: true
      required: [toName, toWallet, amountAudd]
...