# lib/api-zod/src/generated/types/createPaymentRequestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PayMate - AUD on Solana API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePaymentRequestBody {
  toName: string;
  toWallet: string;
  amountAudd: number;
  note?: string | null;
}