# lib/api-zod/src/generated/types/paymentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PayMate - AUD on Solana API
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentRequestStatus } from "./paymentRequestStatus";

export interface PaymentRequest {
  id: string;
  toName: string;
  toWallet: string;
  amountAudd: number;
  note?: string | null;
  status: PaymentRequestStatus;
  createdAt: string;
  paidAt?: string | null;
}